Add patch 1713574: "timer precision fix" by wjp. Fixes vret problems under 64 bit hosts.
Imported-from: https://svn.code.sf.net/p/dosbox/code-0/dosbox/trunk@2868

+++ include/pic.h
@@ -49,7 +49,7 @@ INLINE Bits PIC_MakeCycles(double amount) {
 }
 
 INLINE double PIC_FullIndex(void) {
-	return PIC_Ticks+PIC_TickIndex();
+	return PIC_Ticks+(double)PIC_TickIndex();
 }
 
 void PIC_ActivateIRQ(Bitu irq);
